requirements.txt
discord.py>=2.4.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.0.0,<2.0.0
PyNaCl>=1.5.0,<2.0.0
httpx>=0.27.0
redis>=5.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
boto3>=1.34.0,<2.0.0
stripe>=8.0.0,<9.0.0
aiohttp>=3.9.0,<4.0.0

# Structured logging (JSON file logs)
python-json-logger>=3.0.0

# Prometheus metrics (exposed via /metrics endpoint)
prometheus_client>=0.21.0

# Testing (optional; run: pytest tests/ -v)
pytest>=7.0.0
pytest-asyncio>=0.21.0
aiosqlite>=0.20.0
